# docs/swagger.yaml
basePath: /
definitions:
...
  handlers.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
...
  handlers.RegisterRequest:
    properties:
      address:
        type: string
      email:
        type: string
      password:
        minLength: 3
        type: string
      phone_number:
        type: string
      role:
        description: Optional
        type: string
    required:
    - address
    - email
    - password
    - phone_number
    type: object
...
  handlers.UserResponse:
    properties:
      address:
        type: string
      email:
        type: string
      id:
        type: integer
      phone_number:
        type: string
      role:
        type: string
      token:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is Jakarta Luxury Rent Car service API documentation.
  title: Jakarta Luxury Rent Car API
  version: "1.0"
paths:
...
  /login:
    post:
      consumes:
      - application/json
      description: Login user and return JWT token
      parameters:
      - description: User login request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.UserResponse'
        "400":
          description: Invalid input
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Invalid email or password
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to generate token
          schema:
            additionalProperties: true
            type: object
      summary: User login
      tags:
      - Register and Login
...
  /register:
    post:
      consumes:
      - application/json
      description: Register new user
      parameters:
      - description: User registration request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.UserResponse'
        "400":
          description: Invalid input
          schema:
            additionalProperties: true
            type: object
        "409":
          description: User already exists with this email
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to hash password or register user
          schema:
            additionalProperties: true
            type: object
      summary: Register new user
      tags:
      - Register and Login
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"